+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Links to other posts now open in the app itself
+
 ## [1.59.0] - 2026-01-09
 
 ### Fixed